refactor: split ParamDimVjpAlgorithm into own module, merge misc into _common
- Move ParamDimVjpAlgorithm and its helpers from d_rtrl.py into a new
  param_dim_vjp.py; d_rtrl.py now holds only the D_RTRL alias subclass.
- Migrate the per-algorithm smoke tests out of integration_test.py into each
  algorithm's own test file and delete integration_test.py.
- Merge misc.py helpers into _common.py and fold misc_test.py into
  _common_test.py; update all importers.
