+++ braintrace/_etrace_algorithms/_common.py
@@ -15,11 +15,13 @@
 
 """Shared helpers for the SNN online-learning algorithms."""
 
-from typing import Optional
+from functools import partial
+from typing import Any, Dict, Optional
 
 import brainstate
 import jax
 import jax.numpy as jnp
+import saiunit as u
 
 __all__ = [
     'PresynapticTrace',
@@ -156,3 +158,247 @@ def _resolve_leak(model, explicit: Optional[float]) -> float:
         'Could not resolve the membrane leak factor. Provide `leak=<float>` at '
         'construction, or ensure the model has a state with a `leak` attribute.'
     )
+
+
+def _reset_state_in_a_dict(
+    state_dict: Dict[Any, brainstate.State],
+    batch_size: Optional[int],
+):
+    """
+    Reset the values in a dictionary of states to zero.
+
+    This function iterates over a dictionary of states and resets each state's
+    value to a zero array. The shape of the zero array is determined by the
+    original shape of the state's value and the specified batch size.
+
+    Args:
+        state_dict (Dict[Any, brainstate.State]): A dictionary where keys are any
+            type and values are brainstate.State objects. Each state's value will be
+            reset to a zero array.
+        batch_size (Optional[int]): The size of the batch. If provided, the
+            zero array will include a batch dimension; otherwise, it will not.
+
+    Returns:
+        None: The function modifies the state_dict in place, resetting each
+        state's value to a zero array.
+    """
+    for k, v in state_dict.items():
+        state_dict[k].value = jax.tree.map(partial(_zeros_like_batch_or_not, batch_size), v.value)
+
+
+def _zeros_like_batch_or_not(
+    batch_size: Optional[int],
+    x: jax.Array
+):
+    """
+    Create a zeros array with the same shape and type as the input array,
+    optionally including a batch dimension.
+
+    This function generates a zeros array that matches the shape and data type
+    of the input array `x`. If a batch size is provided, the zeros array will
+    include an additional batch dimension at the beginning.
+
+    Args:
+        batch_size (Optional[int]): The size of the batch. If provided, the
+            zeros array will include a batch dimension. If None, the zeros
+            array will have the same shape as `x`.
+        x (jax.Array): The input array whose shape and data type are used as
+            a reference for creating the zeros array.
+
+    Returns:
+        jax.Array: A zeros array with the same shape and data type as the
+        input array, optionally including a batch dimension if `batch_size`
+        is provided.
+    """
+    if batch_size is not None:
+        assert isinstance(batch_size, int), 'The batch size should be an integer. '
+        return u.math.zeros((batch_size,) + x.shape[1:], x.dtype)
+    else:
+        return u.math.zeros_like(x)
+
+
+def _batched_zeros_like(
+    batch_size: Optional[int],
+    num_state: int,  # the number of hidden states
+    x: jax.Array  # the input array
+):
+    """
+    Create a batched zeros array with the same shape as the input array,
+    extended by the number of hidden states.
+
+    This function generates a zeros array that matches the shape of the
+    input array `x`, with an additional dimension for the number of hidden
+    states. If a batch size is provided, the zeros array will also include
+    a batch dimension.
+
+    Args:
+        batch_size (Optional[int]): The size of the batch. If None, the
+            batch dimension is not included.
+        num_state (int): The number of hidden states, which determines the
+            size of the additional dimension in the zeros array.
+        x (jax.Array): The input array whose shape is used as a reference
+            for creating the zeros array.
+
+    Returns:
+        jax.Array: A zeros array with the same shape as the input array,
+        extended by the number of hidden states, and optionally including
+        a batch dimension.
+    """
+    if batch_size is None:
+        return u.math.zeros((*x.shape, num_state), x.dtype)
+    else:
+        return u.math.zeros((batch_size, *x.shape, num_state), x.dtype)
+
+
+def _sum_dim(xs: jax.Array, axis: int = -1):
+    """
+    Sums the elements along the last dimension of each array in a PyTree.
+
+    This function applies a sum operation along the last dimension of each array
+    within a PyTree structure. It is useful for reducing the dimensionality of
+    arrays by aggregating values along the specified axis.
+
+    Args:
+        xs (jax.Array): A PyTree of arrays where each array will have its last
+                        dimension summed.
+
+    Returns:
+        jax.Array: A PyTree with the same structure as the input, where each array
+                   has been reduced by summing over its last dimension.
+    """
+    return jax.tree.map(lambda x: u.math.sum(x, axis=axis), xs)
+
+
+def _unit_safe_add(a, b):
+    """Add two leaves, stripping units only when one side has units and the other does not.
+
+    Gradient contributions for the same weight may come from paths that
+    preserve physical units (e.g. VJP through the original jaxpr) and paths
+    that already strip them (e.g. ETP ``xy_to_dw`` rules). When the two sides
+    disagree on unit representation, both are reduced to plain arrays before
+    adding; otherwise units are preserved.
+    """
+    a_is_q = isinstance(a, u.Quantity)
+    b_is_q = isinstance(b, u.Quantity)
+    if a_is_q != b_is_q:
+        a = u.get_mantissa(a) if a_is_q else a
+        b = u.get_mantissa(b) if b_is_q else b
+    return u.math.add(a, b)
+
+
+def _extract_leaf(pytree_val: brainstate.typing.PyTree, leaf_idx: int):
+    """Return the leaf at ``leaf_idx`` in ``jax.tree.leaves(pytree_val)``.
+
+    Bare arrays (treedef with a single leaf) return the array unchanged.
+    Raises ``IndexError`` if ``leaf_idx`` is outside ``len(leaves)``.
+    """
+    leaves = jax.tree.leaves(pytree_val)
+    if not leaves:
+        return pytree_val
+    if leaf_idx < 0 or leaf_idx >= len(leaves):
+        raise IndexError(
+            f'leaf_idx {leaf_idx} out of range for pytree with {len(leaves)} leaves'
+        )
+    return leaves[leaf_idx]
+
+
+def _wrap_leaves_as_pytree(
+    reference_pytree: brainstate.typing.PyTree,
+    leaf_grads: Dict[int, jax.Array],
+):
+    """Build a pytree matching ``reference_pytree`` with ``leaf_grads``
+    inserted at the given leaf indices; any other leaf is zero-filled.
+
+    When the reference is a bare array, ``leaf_grads`` must contain at most one
+    entry at index 0 and that value is returned directly (no wrapping).
+
+    Raises ``IndexError`` if any supplied index is outside
+    ``len(jax.tree.leaves(reference_pytree))``.
+    """
+    ref_treedef = jax.tree.structure(reference_pytree)
+    # Bare-array fast path.
+    # jax's PyTreeDef stubs omit num_leaves and __eq__; both are valid at runtime.
+    if ref_treedef.num_leaves <= 1 and ref_treedef == jax.tree.structure(0):  # type: ignore[attr-defined, operator]
+        if 0 in leaf_grads:
+            return leaf_grads[0]
+        return u.math.zeros_like(reference_pytree)
+    leaves = jax.tree.leaves(reference_pytree)
+    n = len(leaves)
+    for idx in leaf_grads:
+        if idx < 0 or idx >= n:
+            raise IndexError(
+                f'leaf_idx {idx} out of range for pytree with {n} leaves'
+            )
+    new_leaves = [
+        leaf_grads[i] if i in leaf_grads else u.math.zeros_like(leaf)
+        for i, leaf in enumerate(leaves)
+    ]
+    return jax.tree.unflatten(ref_treedef, new_leaves)
+
+
+def _route_grads_by_path(
+    relation,
+    per_key_grads: Dict[str, jax.Array],
+    weight_vals: Dict[Any, brainstate.typing.PyTree],
+    target_dict: Dict[Any, brainstate.typing.PyTree],
+) -> None:
+    """Route per-key gradients from a dict-API rule into per-path pytrees.
+
+    Both D-RTRL and ES-D-RTRL share this bookkeeping: for each key in
+    ``per_key_grads`` (returned by ``xy_to_dw`` or ``yw_to_w``), look up the
+    owning ``ParamState`` path and the leaf index from ``relation``, accumulate
+    into ``per_path``, then wrap with ``_wrap_leaves_as_pytree`` and merge into
+    ``target_dict`` via ``_update_dict``.
+
+    Args:
+        relation: HiddenParamOpRelation — provides ``trainable_paths`` and
+            ``trainable_leaf_indices``.
+        per_key_grads: Dict[str, Array] — gradient contributions keyed by
+            trainable invar name (e.g. ``'weight'``, ``'lora_b'``).
+        weight_vals: Dict[Path, PyTree] — current ParamState pytree values;
+            used as the structure template for ``_wrap_leaves_as_pytree``.
+        target_dict: Dict[Path, PyTree] — accumulation target, modified in place.
+    """
+    per_path: Dict[Any, Dict[int, jax.Array]] = {}
+    for key, grad in per_key_grads.items():
+        path = relation.trainable_paths[key]
+        leaf_idx = relation.trainable_leaf_indices[key]
+        per_path.setdefault(path, {})[leaf_idx] = grad
+    for path, leaf_to_grad in per_path.items():
+        wrapped = _wrap_leaves_as_pytree(weight_vals[path], leaf_to_grad)
+        _update_dict(target_dict, path, wrapped)
+
+
+def _update_dict(
+    the_dict: Dict,
+    key: Any,
+    value: brainstate.typing.PyTree,
+    error_when_no_key: Optional[bool] = False
+):
+    """Update the dictionary.
+
+    If the key exists, then add the value to the existing value.
+    Otherwise, create a new key-value pair.
+
+    Args:
+      the_dict: The dictionary.
+      key: The key.
+      value: The value.
+      error_when_no_key: bool, whether to raise an error when the key does not exist.
+
+    """
+    if key not in the_dict:
+        if error_when_no_key:
+            raise ValueError(f'The key {key} does not exist in the dictionary. ')
+        the_dict[key] = value
+    else:
+        old_value = the_dict[key]
+        if old_value is None:
+            the_dict[key] = value
+        else:
+            the_dict[key] = jax.tree.map(
+                _unit_safe_add,
+                old_value,
+                value,
+                is_leaf=lambda x: isinstance(x, u.Quantity)
+            )
