+++ braintrace/_etrace_algorithms/cross_check_test.py
@@ -27,7 +27,7 @@ import jax.numpy as jnp
 
 import braintrace
 from braintrace._etrace_algorithms import EProp
-from braintrace._etrace_algorithms.d_rtrl import ParamDimVjpAlgorithm
+from braintrace._etrace_algorithms.param_dim_vjp import ParamDimVjpAlgorithm
 
 
 def _net():
