+++ braintrace/_etrace_algorithms/e_prop_test.py
@@ -71,7 +71,7 @@ class TestEPropConstruction(unittest.TestCase):
 class TestEPropKappaApplied(unittest.TestCase):
     def test_kappa_zero_matches_d_rtrl(self):
         """κ=0 must reproduce D_RTRL gradients bit-for-bit on the same model."""
-        from braintrace._etrace_algorithms.d_rtrl import ParamDimVjpAlgorithm
+        from braintrace._etrace_algorithms.param_dim_vjp import ParamDimVjpAlgorithm
 
         def compute(algo_cls, **extra):
             net = _lsnn_like()
@@ -137,3 +137,56 @@ class TestEPropRandomFeedback(unittest.TestCase):
         g_sym = compute('symmetric')
         g_rnd = compute('random', random_feedback_key=jax.random.PRNGKey(123))
         assert not jnp.allclose(g_sym, g_rnd, atol=1e-4)
+
+
+class FakeLIF(brainstate.HiddenState):
+    def __init__(self, iv, leak):
+        super().__init__(iv)
+        self.leak = leak
+
+
+def _toy_net():
+    class Net(brainstate.nn.Module):
+        def __init__(self):
+            super().__init__()
+            self.w = brainstate.ParamState(
+                0.1 * jax.random.normal(jax.random.PRNGKey(0), (3, 3))
+            )
+            self.v = FakeLIF(jnp.zeros((1, 3)), leak=0.9)
+
+        def update(self, x):
+            self.v.value = jax.nn.tanh(
+                0.9 * self.v.value + braintrace.matmul(x, self.w.value)
+            )
+            return self.v.value
+
+    net = Net()
+    brainstate.nn.init_all_states(net, batch_size=1)
+    return net
+
+
+def _run(algo, n_steps=10, lr=0.05, y_target=None, pass_y=False):
+    x = jnp.ones((1, 3))
+    algo.compile_graph(x)
+    algo.init_etrace_state()
+
+    losses = []
+    for _ in range(n_steps):
+        def loss_fn(x_):
+            out = algo.update(x_, y_target=y_target) if pass_y else algo.update(x_)
+            target = jnp.ones_like(out)
+            return ((out - target) ** 2).mean()
+
+        grads, loss_val = brainstate.augment.grad(
+            loss_fn, algo.param_states, return_value=True
+        )(x)
+        for path, st in algo.param_states.items():
+            st.value = st.value - lr * grads[path]
+        losses.append(float(loss_val))
+    return losses
+
+
+class TestSmokeLossDecreases(unittest.TestCase):
+    def test_eprop(self):
+        losses = _run(EProp(_toy_net()))
+        assert losses[-1] < losses[0]
