+++ braintrace/_etrace_algorithms/ostl.py
@@ -41,7 +41,7 @@ from typing import Optional
 
 import brainstate
 
-from .d_rtrl import ParamDimVjpAlgorithm
+from .param_dim_vjp import ParamDimVjpAlgorithm
 from .pp_prop import pp_prop
 
 __all__ = ['OSTL', 'OSTLRecurrent', 'OSTLFeedforward']
