+++ braintrace/_etrace_algorithms/ottt.py
@@ -31,8 +31,7 @@ import brainstate
 import jax.numpy as jnp
 
 from braintrace._etrace_op import is_batched_primitive
-from ._common import PresynapticTrace, _resolve_leak
-from .misc import _route_grads_by_path, _update_dict
+from ._common import PresynapticTrace, _resolve_leak, _route_grads_by_path, _update_dict
 from .vjp_base import ETraceVjpAlgorithm
 
 __all__ = ['OTTT']
